refactor: complete LOW-severity code quality fixes
- TxcActivitySource.CurrentOperationId: consolidated Activity.Current?.TraceId pattern (5 call sites)
- ExecutionLogService: extracted from McpToolResultFactory (SRP — builder vs query/resource)
- GlobalConfig.Log.ApplyAsEnvironmentDefaults(): wires stored log settings to TxcLoggerFactory
- CliOutputTests: handle Debug builds gracefully (support info requires Release telemetry)
