+++ src/TALXIS.CLI.MCP/McpToolResultFactory.cs
@@ -1,6 +1,4 @@
-using System.Diagnostics;
 using System.Text.Json;
-using ModelContextProtocol;
 using ModelContextProtocol.Protocol;
 using TALXIS.CLI.Abstractions;
 using TALXIS.CLI.Core;
@@ -9,7 +7,8 @@ using TALXIS.CLI.Logging;
 namespace TALXIS.CLI.MCP;
 
 /// <summary>
-/// Builds MCP tool results and exposes execution logs as fetchable resources.
+/// Builds MCP <see cref="CallToolResult"/> responses from CLI subprocess outcomes.
+/// Log query and resource operations are handled by <see cref="ExecutionLogService"/>.
 /// </summary>
 internal sealed class McpToolResultFactory
 {
@@ -29,7 +28,7 @@ internal sealed class McpToolResultFactory
     {
         // Capture operation ID while Activity is still active — by the time
         // BuildFailureResult runs the Activity scope may have ended.
-        var operationId = Activity.Current?.TraceId.ToHexString();
+        var operationId = TxcActivitySource.CurrentOperationId;
 
         // Store execution log for every run — clients may need to troubleshoot
         // unexpected output even from successful executions.
@@ -76,7 +75,7 @@ internal sealed class McpToolResultFactory
     public CallToolResult BuildExceptionResult(string toolName, Exception exception)
     {
         // Capture operation ID while Activity is still active.
-        var operationId = Activity.Current?.TraceId.ToHexString();
+        var operationId = TxcActivitySource.CurrentOperationId;
 
         // Surface the innermost exception message — for wrapped exceptions
         // (e.g. HttpRequestException → SocketException), the root cause is
@@ -115,105 +114,6 @@ internal sealed class McpToolResultFactory
         {
             return null;
         }
-    }
-
-    public List<Resource> BuildResources()
-    {
-        return _toolLogStore.ListAll().Select(e => new Resource
-        {
-            Uri = e.Uri,
-            Name = $"Execution log: {e.Entry.ToolName}",
-            Description = BuildResourceDescription(e.Entry),
-            MimeType = "application/json"
-        }).ToList();
-    }
-
-    public ReadResourceResult ReadResource(string uri)
-    {
-        if (!_toolLogStore.TryGet(uri, out var entry) || entry is null)
-        {
-            throw new McpException($"Resource not found: {uri}");
-        }
-
-        return new ReadResourceResult
-        {
-            Contents = [new TextResourceContents { Uri = uri, MimeType = "application/json", Text = entry.ToJson() }]
-        };
-    }
-
-    public CallToolResult BuildExecutionLogResult(string uri, string? level = null,
-        string? category = null, string? search = null, int skip = 0, int take = 50)
-    {
-        if (!_toolLogStore.TryGet(uri, out var entry) || entry is null)
-        {
-            return new CallToolResult
-            {
-                IsError = true,
-                Content = [new TextContentBlock { Text = $"Execution log not found for '{uri}'." }]
-            };
-        }
-
-        var allEntries = entry.LogEntries;
-        var filtered = FilterLogEntries(allEntries, level, category, search);
-        var paged = filtered.Skip(skip).Take(take).ToList();
-
-        var result = new
-        {
-            entry.ToolName,
-            entry.ExitCode,
-            entry.Summary,
-            TotalEntries = allEntries.Count,
-            FilteredCount = filtered.Count,
-            Skip = skip,
-            Take = take,
-            LogEntries = paged
-        };
-
-        return new CallToolResult
-        {
-            Content = [new TextContentBlock
-            {
-                Text = JsonSerializer.Serialize(result, TxcOutputJsonOptions.Default)
-            }]
-        };
-    }
-
-    private static readonly string[] LogLevelOrder =
-        ["Trace", "Debug", "Information", "Warning", "Error", "Critical"];
-
-    private static IReadOnlyList<RedactedLogEntry> FilterLogEntries(
-        IReadOnlyList<RedactedLogEntry> entries, string? level, string? category, string? search)
-    {
-        IEnumerable<RedactedLogEntry> result = entries;
-
-        if (!string.IsNullOrWhiteSpace(level))
-        {
-            int minIndex = Array.IndexOf(LogLevelOrder, level);
-            if (minIndex >= 0)
-            {
-                result = result.Where(e =>
-                {
-                    int idx = Array.IndexOf(LogLevelOrder, e.Level);
-                    return idx >= minIndex;
-                });
-            }
-        }
-
-        if (!string.IsNullOrWhiteSpace(category))
-        {
-            result = result.Where(e =>
-                e.Category.Contains(category, StringComparison.OrdinalIgnoreCase));
-        }
-
-        if (!string.IsNullOrWhiteSpace(search))
-        {
-            result = result.Where(e =>
-                e.Message.Contains(search, StringComparison.OrdinalIgnoreCase) ||
-                (e.Data != null && e.Data.Values.Any(v =>
-                    v?.ToString()?.Contains(search, StringComparison.OrdinalIgnoreCase) == true)));
-        }
-
-        return result.ToList();
     }
 
     private CallToolResult BuildFailureResult(string toolName, string summary,
@@ -273,20 +173,6 @@ internal sealed class McpToolResultFactory
         return !string.IsNullOrWhiteSpace(firstErrorLine)
             ? firstErrorLine
             : $"Tool '{toolName}' failed with exit code {exitCode}.";
-    }
-
-    private static string BuildResourceDescription(ToolLogStore.LogEntry entry)
-    {
-        var singleLineSummary = entry.Summary
-            .ReplaceLineEndings(" ")
-            .Trim();
-
-        if (singleLineSummary.Length > 160)
-        {
-            singleLineSummary = singleLineSummary[..157] + "...";
-        }
-
-        return $"{singleLineSummary} Use resources/read to retrieve detailed diagnostics.";
     }
 
 }
